% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mbr.R
\name{geom_mbr}
\alias{geom_mbr}
\title{Title}
\usage{
geom_mbr(
  mapping = NULL,
  data = NULL,
  stat = "MBR",
  position = "identity",
  ...,
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{na.rm}{}
}
\value{

}
\description{
Title
}
